[ADM-3763] add modal confirmation and copies

+++ src/app/profile/profile-pro/profile-pro-billing/profile-pro-billing.component.ts
@@ -4,6 +4,8 @@ import { PaymentService } from '../../../core/payments/payment.service';
 import { ErrorsService } from '../../../core/errors/errors.service';
 import { UUID } from 'angular2-uuid';
 import { BillingInfoResponse } from '../../../core/payments/payment.interface';
+import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
+import { DeleteInfoConfirmationModalComponent } from './delete-info-confirmation-modal/delete-info-confirmation-modal.component';
 
 @Component({
   selector: 'tsl-profile-pro-billing',
@@ -17,7 +19,8 @@ export class ProfileProBillingComponent implements OnInit {
 
   constructor(private fb: FormBuilder,
               private paymentService: PaymentService,
-              private errorsService: ErrorsService) {
+              private errorsService: ErrorsService,
+              private modalService: NgbModal) {
     this.billingForm = fb.group({
       cif: ['', [Validators.required]],
       city: ['', [Validators.required]],
@@ -49,6 +52,7 @@ export class ProfileProBillingComponent implements OnInit {
     if (this.billingForm.valid) {
       this.paymentService.updateBillingInfo(this.billingForm.value).subscribe(() => {
         this.errorsService.i18nSuccess('userEdited');
+        this.isNewBillingInfoForm = false;
       }, (response: any) => {
         this.errorsService.show(response);
       });
@@ -63,12 +67,16 @@ export class ProfileProBillingComponent implements OnInit {
   }
 
   public deleteBillingInfo() {
-    this.paymentService.deleteBillingInfo(this.billingForm.value.id).subscribe(() => {
-      this.errorsService.i18nSuccess('deleteBillingInfoSuccess');
-      this.billingForm.reset();
-      this.isNewBillingInfoForm = true;
-    }, () => {
-      this.errorsService.i18nError('deleteBillingInfoError');
+    this.modalService.open(DeleteInfoConfirmationModalComponent).result.then((result: boolean) => {
+      if (result) {
+        this.paymentService.deleteBillingInfo(this.billingForm.value.id).subscribe(() => {
+          this.errorsService.i18nSuccess('deleteBillingInfoSuccess');
+          this.billingForm.reset();
+          this.isNewBillingInfoForm = true;
+        }, () => {
+          this.errorsService.i18nError('deleteBillingInfoError');
+        });
+      }
     });
   }
 }
